Move SelfLaunchTimeoutPolicy.kt to where it belongs
